Expand the PHP short tags to long tags.
Modernize and ensure compatibility across servers. Really the short tag
is dead now, all major projects are outlawing this style.

  http://php.net/manual/en/language.basic-syntax.phptags.php

  http://pear.php.net/manual/en/standards.tags.php

Always use <?php ?> to delimit PHP code, not the <? ?> shorthand.
This is required for PEAR compliance and is also the most portable
way to include PHP code on differing operating systems and setups.